+++ src/main/java/subway/persistence/entity/StationEntity.java
@@ -1,0 +1,23 @@
+package subway.persistence.entity;
+
+public class StationEntity {
+    private final Long id;
+    private final String name;
+
+    public StationEntity(final String name) {
+        this(null, name);
+    }
+
+    public StationEntity(final Long id, final String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
