+++ src/main/java/subway/ui/dto/request/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.ui.dto.request;
 
 public class LineRequest {
     private String name;
