+++ src/main/java/subway/ui/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.ui.dto.request;
 
 public class StationRequest {
     private String name;
