+++ src/main/java/subway/ui/dto/response/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.ui.dto.response;
 
 import subway.domain.Line;
